[Deps] Updated to Rails 4.2 (w/ responders)
* Updated to Rails 4.2
* Refactored custom responders out of config
* Not using Rails API anymore
* Dropping logstasher repo source
